Fix Sollbuchung Export
