+++ src/de/jost_net/JVerein/gui/action/MitgliedskontoExportAction.java
@@ -24,7 +24,6 @@ import de.jost_net.JVerein.gui.view.DokumentationUtil;
 import de.jost_net.JVerein.io.Exporter;
 import de.jost_net.JVerein.io.IORegistry;
 import de.jost_net.JVerein.io.MitgliedskontoExport;
-import de.jost_net.JVerein.rmi.Mitglied;
 import de.jost_net.JVerein.rmi.Mitgliedskonto;
 import de.willuhn.jameica.gui.Action;
 import de.willuhn.jameica.gui.GUI;
@@ -36,30 +35,10 @@ public class MitgliedskontoExportAction implements Action
 {
   private EXPORT_TYP exportTyp;
 
-  private Mitglied selectedMitglied;
 
   public MitgliedskontoExportAction(EXPORT_TYP exportTyp, Object currentObject)
   {
-    try
-    {
-      this.exportTyp = exportTyp;
-      if (null != currentObject)
-      {
-        if (currentObject instanceof Mitglied)
-        {
-          selectedMitglied = (Mitglied) currentObject;
-        }
-        else if (currentObject instanceof Mitgliedskonto)
-        {
-          Mitgliedskonto konto = (Mitgliedskonto) currentObject;
-          selectedMitglied = konto.getMitglied();
-        }
-      }
-    }
-    catch (RemoteException e)
-    {
-      Logger.error("Export kann nicht initialisiert werden", e);
-    }
+    this.exportTyp = exportTyp;
   }
 
   /**
@@ -88,7 +67,6 @@ public class MitgliedskontoExportAction implements Action
     {
       Logger.error("Fehler", e);
       GUI.getStatusBar().setErrorText(
-
       "Fehler beim exportieren der Sollbuchungen");
     }
   }
@@ -110,12 +88,12 @@ public class MitgliedskontoExportAction implements Action
     }
   }
 
-  private Object[] gibSuchGrenzen(Object context) throws ApplicationException
+  private Object[] gibSuchGrenzen(Object context) throws ApplicationException, RemoteException
   {
     if (context instanceof MitgliedskontoControl)
     {
       MitgliedskontoControl control = (MitgliedskontoControl) context;
-      return control.getCVSExportGrenzen(selectedMitglied);
+      return control.getCVSExportGrenzen();
     }
     throw new ApplicationException(
         "Dieser Export wurde aus dem falschen Context aufgerufen!");
