+++ src/de/jost_net/JVerein/gui/control/MitgliedskontoControl.java
@@ -16,10 +16,7 @@
  **********************************************************************/
 package de.jost_net.JVerein.gui.control;
 
-import java.math.BigDecimal;
 import java.rmi.RemoteException;
-import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
@@ -32,10 +29,10 @@ import org.eclipse.swt.widgets.TreeItem;
 
 import de.jost_net.JVerein.Einstellungen;
 import de.jost_net.JVerein.Messaging.MitgliedskontoMessage;
+import de.jost_net.JVerein.Queries.SollbuchungQuery;
 import de.jost_net.JVerein.gui.formatter.ZahlungswegFormatter;
 import de.jost_net.JVerein.gui.input.BuchungsartInput;
 import de.jost_net.JVerein.gui.input.BuchungsklasseInput;
-import de.jost_net.JVerein.gui.input.MailAuswertungInput;
 import de.jost_net.JVerein.gui.input.MitgliedInput;
 import de.jost_net.JVerein.gui.input.BuchungsartInput.buchungsarttyp;
 import de.jost_net.JVerein.gui.menu.MitgliedskontoMenu;
@@ -51,11 +48,7 @@ import de.jost_net.JVerein.rmi.Mitglied;
 import de.jost_net.JVerein.rmi.Mitgliedskonto;
 import de.jost_net.JVerein.util.JVDateFormatTTMMJJJJ;
 import de.willuhn.datasource.GenericIterator;
-import de.willuhn.datasource.GenericObject;
-import de.willuhn.datasource.pseudo.PseudoIterator;
 import de.willuhn.datasource.rmi.DBIterator;
-import de.willuhn.datasource.rmi.DBService;
-import de.willuhn.datasource.rmi.ResultSetExtractor;
 import de.willuhn.jameica.gui.AbstractView;
 import de.willuhn.jameica.gui.Action;
 import de.willuhn.jameica.gui.GUI;
@@ -518,7 +511,7 @@ public class MitgliedskontoControl extends DruckMailControl
     this.action = action;
     this.umwandeln = umwandeln;
     @SuppressWarnings("rawtypes")
-    GenericIterator mitgliedskonten = getMitgliedskontoIterator(umwandeln);
+    GenericIterator mitgliedskonten = new SollbuchungQuery(this, umwandeln, null).get();
     if (mitgliedskontoList == null)
     {
       mitgliedskontoList = new SollbuchungListTablePart(mitgliedskonten, action);
@@ -626,11 +619,10 @@ public class MitgliedskontoControl extends DruckMailControl
     return mitglieder;
   }
 
-  
   public void refreshMitgliedkonto1() throws RemoteException
   {
     @SuppressWarnings("rawtypes")
-    GenericIterator mitgliedskonten = getMitgliedskontoIterator(umwandeln);
+    GenericIterator mitgliedskonten = new SollbuchungQuery(this, umwandeln, null).get();
     mitgliedskontoList.removeAll();
     while (mitgliedskonten.hasNext())
     {
@@ -638,379 +630,6 @@ public class MitgliedskontoControl extends DruckMailControl
     }
   }
   
-  @SuppressWarnings("rawtypes")
-  public GenericIterator getMitgliedskontoIterator(boolean umwandeln) throws RemoteException
-  {
-    this.umwandeln = umwandeln;
-    Date d1 = null;
-    java.sql.Date vd = null;
-    java.sql.Date bd = null;
-    if (datumvon != null)
-    {
-      d1 = (Date) datumvon.getValue();
-      if (d1 != null)
-      {
-        vd = new java.sql.Date(d1.getTime());
-      }
-    }
-    if (datumbis != null)
-    {
-      d1 = (Date) datumbis.getValue();
-      if (d1 != null)
-      {
-        bd = new java.sql.Date(d1.getTime());
-      }
-    }
-    
-    DIFFERENZ diff = DIFFERENZ.EGAL;
-    if (differenz != null)
-    {
-      diff = (DIFFERENZ) differenz.getValue();
-    }
-    
-    boolean kein_name = suchname == null || suchname.getValue() == null || 
-        ((String) suchname.getValue()).isEmpty();
-    boolean ein_name = suchname != null && suchname.getValue() != null && 
-        !((String) suchname.getValue()).isEmpty();
-    boolean keine_email = mailAuswahl == null || (Integer) mailAuswahl.getValue() == 
-        MailAuswertungInput.ALLE;
-    boolean filter_email = mailAuswahl != null && !((Integer) mailAuswahl.getValue() == 
-        MailAuswertungInput.ALLE);
-    
-    // Falls kein Name, kein Mailfilter und keine Differenz dann alles lesen
-    if (kein_name && keine_email &&  diff == DIFFERENZ.EGAL)
-    {
-      DBIterator<Mitgliedskonto> sollbuchungen = Einstellungen.getDBService()
-          .createList(Mitgliedskonto.class);
-      if (vd != null)
-      {
-      sollbuchungen.addFilter("mitgliedskonto.datum >= ? ",
-          new Object[] { vd });
-      }
-      if (bd != null)
-      {
-      sollbuchungen.addFilter("mitgliedskonto.datum <= ? ",
-          new Object[] { bd });
-      }
-      if (ohneabbucher != null && (Boolean) ohneabbucher.getValue())
-      {
-        sollbuchungen.addFilter("mitgliedskonto.zahlungsweg <> ?", 
-            Zahlungsweg.BASISLASTSCHRIFT);
-      }
-      return sollbuchungen;
-    }
-    
-    // Falls ein Name oder Mailfilter aber keine Differenz dann alles des Mitglieds lesen
-    if ((ein_name || filter_email)  && diff == DIFFERENZ.EGAL)
-    {
-      DBIterator<Mitgliedskonto> sollbuchungen = Einstellungen.getDBService()
-          .createList(Mitgliedskonto.class);
-      
-      if ((!umwandeln && ein_name) || filter_email)
-      {
-        sollbuchungen.join("mitglied");
-        sollbuchungen.addFilter("mitglied.id = mitgliedskonto.mitglied");
-      }
-      
-      if (!umwandeln && ein_name)
-      {
-        // Der Name kann so verwendet werden ohne Umwandeln der Umlaute
-        String name = (String) suchname.getValue();
-        sollbuchungen.addFilter("((lower(mitglied.name) like ?)"
-            + " OR (lower(mitglied.vorname) like ?))",
-            new Object[] {name.toLowerCase() + "%", name.toLowerCase() + "%"});
-      }
-      else if (umwandeln && ein_name)
-      {
-        // Der Name muss umgewandelt werden, es kann mehrere Matches geben
-        ArrayList<BigDecimal> namenids = getNamenIds();
-        if (namenids != null)
-        {
-          int anzahl = namenids.size();
-          String querystring = null;
-          
-          for (int i = 1; i <= anzahl; i++)
-          {
-            if (anzahl == 1)
-            {
-              querystring = "(mitgliedskonto.mitglied = ?) ";
-            }
-            else if (i == 1)
-            {
-              querystring =  "((mitgliedskonto.mitglied = ?) OR ";
-            }
-            else if (i < anzahl)
-            {
-              querystring =  querystring + "(mitgliedskonto.mitglied = ?) OR ";
-            }
-            else if (i == anzahl)
-            {
-              querystring =  querystring + "(mitgliedskonto.mitglied = ?)) ";
-            }
-          }
-          sollbuchungen.addFilter(querystring, namenids.toArray() );
-        }
-      }
-
-      if (vd != null)
-      {
-        sollbuchungen.addFilter("(mitgliedskonto.datum >= ?) ",
-            new Object[] { vd });
-      }
-      if (bd != null)
-      {
-        sollbuchungen.addFilter("(mitgliedskonto.datum <= ?) ",
-            new Object[] { bd });
-      }
-      if (ohneabbucher != null && (Boolean) ohneabbucher.getValue())
-      {
-        sollbuchungen.addFilter("mitgliedskonto.zahlungsweg <> ?", 
-            Zahlungsweg.BASISLASTSCHRIFT);
-      }
-      if (filter_email)
-      {
-        int mailauswahl = (Integer) mailAuswahl.getValue();
-        if (mailauswahl == MailAuswertungInput.OHNE)
-        {
-          sollbuchungen.addFilter("(email is null or length(email) = 0)");
-        }
-        if (mailauswahl == MailAuswertungInput.MIT)
-        {
-          sollbuchungen.addFilter("(email is  not null and length(email) > 0)");
-        }
-      }
-      return sollbuchungen;
-    }
-    
-    // Eine Differenz ist ausgewählt
-    final DBService service = Einstellungen.getDBService();
-    String sql = "SELECT  mitgliedskonto.id, mitglied.name, mitglied.vorname, "
-        + " mitgliedskonto.betrag, sum(buchung.betrag) FROM mitgliedskonto "
-        + "JOIN mitglied on (mitgliedskonto.mitglied = mitglied.id) "
-        + "LEFT JOIN buchung on mitgliedskonto.id = buchung.mitgliedskonto ";
-    String where = "";
-    ArrayList<Object> param = new ArrayList<>();
-    if (suchname != null && suchname.getValue() != null && 
-        !((String) suchname.getValue()).isEmpty() && umwandeln == false)
-    {
-      // Der Name kann so verwendet werden ohne Umwandeln der Umlaute
-      String tmpSuchname = (String) suchname.getValue();
-      where += (where.length() > 0 ? "and " : "")
-          + "((lower(mitglied.name) like ?) OR (lower(mitglied.vorname) like ?)) ";
-      param.add(tmpSuchname.toLowerCase() + "%");
-      param.add(tmpSuchname.toLowerCase() + "%");
-    }
-    else if (suchname != null && suchname.getValue() != null && 
-        !((String) suchname.getValue()).isEmpty() && umwandeln == true)
-    {
-      // Der Name muss umgewandelt werden, es kann mehrere Matches geben
-      ArrayList<BigDecimal> namenids = getNamenIds();
-      if (namenids != null)
-      {
-        int count = 0;
-        int anzahl = namenids.size();
-        for (BigDecimal id: namenids)
-        {
-          count++;
-          if (anzahl == 1)
-          {
-          where += (where.length() > 0 ? "and " : "")
-              + "mitgliedskonto.mitglied = ? ";
-          }
-          else if (count == 1)
-          {
-            where += (where.length() > 0 ? "and " : "")
-                + "(mitgliedskonto.mitglied = ? ";
-          }
-          else if (count < anzahl)
-          {
-            where += " OR mitgliedskonto.mitglied = ? ";
-          }
-          else if (count == anzahl)
-          {
-            where += " OR mitgliedskonto.mitglied = ?) ";
-          }
-          param.add(id);
-        }
-      }
-    }
-    if (vd != null)
-    {
-      where += (where.length() > 0 ? "and " : "")
-          + "mitgliedskonto.datum >= ? ";
-      param.add(vd);
-    }
-    if (bd != null)
-    {
-      where += (where.length() > 0 ? "and " : "")
-          + "mitgliedskonto.datum <= ? ";
-      param.add(bd);
-    }
-    if (ohneabbucher != null && (Boolean) ohneabbucher.getValue())
-    {
-      where += (where.length() > 0 ? "and " : "")
-          +"mitgliedskonto.zahlungsweg <> ?"; 
-      param.add(Zahlungsweg.BASISLASTSCHRIFT);
-    }
-    if (filter_email)
-    {
-      int mailauswahl = (Integer) mailAuswahl.getValue();
-      if (mailauswahl == MailAuswertungInput.OHNE)
-      {
-        where += (where.length() > 0 ? "and " : "")
-            + "(email is null or length(email) = 0)";
-      }
-      if (mailauswahl == MailAuswertungInput.MIT)
-      {
-        where += (where.length() > 0 ? "and " : "")
-            + "(email is not null and length(email) > 0)";
-      }
-    }
-    
-    if (where.length() > 0)
-    {
-      sql += "WHERE " + where;
-    }
-    sql += "group by mitgliedskonto.id ";
-
-    if (DIFFERENZ.FEHLBETRAG == diff)
-    {
-      sql += "having sum(buchung.betrag) < mitgliedskonto.betrag or "
-          + "(sum(buchung.betrag) is null and mitgliedskonto.betrag > 0) ";
-    }
-    if (DIFFERENZ.UEBERZAHLUNG == diff)
-    {
-      sql += "having sum(buchung.betrag) > mitgliedskonto.betrag ";
-    }
-    sql += "order by mitglied.name, mitglied.vorname, mitgliedskonto.datum desc";
-    @SuppressWarnings("unchecked")
-    ArrayList<Mitgliedskonto> mitgliedskonten = (ArrayList<Mitgliedskonto>) service.execute(sql,
-        param.toArray(), new ResultSetExtractor()
-    {
-      @Override
-      public Object extract(ResultSet rs)
-          throws RemoteException, SQLException
-      {
-        ArrayList<Mitgliedskonto> list = new ArrayList<>();
-        while (rs.next())
-        {
-          list.add(
-            (Mitgliedskonto) service.createObject(Mitgliedskonto.class, rs.getString(1)));
-        }
-        return list;
-      }
-    });
-    
-    return PseudoIterator.fromArray(
-        mitgliedskonten.toArray(new GenericObject[mitgliedskonten.size()]));
-  }
-  
-  private ArrayList<BigDecimal> getNamenIds() throws RemoteException
-  {
-    DBService service = Einstellungen.getDBService();
-    String sql = "SELECT  mitglied.id, mitglied.name, mitglied.vorname from mitglied";
-    
-    @SuppressWarnings("unchecked")
-    ArrayList<BigDecimal> mitgliedids = (ArrayList<BigDecimal>) service.execute(sql,
-        new Object[] { }, new ResultSetExtractor()
-        {
-          @Override
-          public Object extract(ResultSet rs)
-              throws RemoteException, SQLException
-          {
-            ArrayList<BigDecimal> ergebnis = new ArrayList<>();
-
-            // In case the text search input is used, we calculate
-            // an "equality" score for each Mitglied entry. 
-            // Only the entries with
-            // score == maxScore will be shown.
-            Integer maxScore = 0;
-            int count = 0;
-            String name = reduceWord((String) suchname.getValue());
-            BigDecimal mgid = null;
-            String nachname = null;
-            String vorname = null;
-            while (rs.next())
-            {
-              count++;
-              // Nur die ids der Mitglieder speichern
-              mgid = rs.getBigDecimal(1);
-
-              StringTokenizer tok = new StringTokenizer(name, " ,-");
-              Integer score = 0;
-              nachname = reduceWord(rs.getString(2));
-              vorname = reduceWord(rs.getString(3));                
-              while (tok.hasMoreElements())
-              {
-                String nextToken = tok.nextToken();
-                if (nextToken.length() > 2)
-                {
-                  score += scoreWord(nextToken, nachname);
-                  score += scoreWord(nextToken, vorname);
-                }
-              }
-
-              if (maxScore < score)
-              {
-                maxScore = score;
-                // We found a Mitgliedskonto matching with a higher equality
-                // score, so we drop all previous matches, because they were
-                // less equal.
-                ergebnis.clear();
-              }
-              else if (maxScore > score)
-              {
-                // This match is worse, so skip it.
-                continue;
-              }
-              ergebnis.add(mgid);
-            }
-            if (ergebnis.size() != count)
-            {
-              return ergebnis;
-            }
-            else
-            {
-              // Kein Match
-              return null;
-            }
-          }
-        });
-    return mitgliedids;
-  }
-
-  public Integer scoreWord(String word, String in)
-  {
-    Integer wordScore = 0;
-    StringTokenizer tok = new StringTokenizer(in, " ,-");
-
-    while (tok.hasMoreElements())
-    {
-      String nextToken = tok.nextToken();
-
-      // Full match is twice worth
-      if (nextToken.equals(word))
-      {
-        wordScore += 2;
-      }
-      else if (nextToken.contains(word))
-      {
-        wordScore += 1;
-      }
-    }
-
-    return wordScore;
-  }
-
-  public String reduceWord(String word)
-  {
-    // We replace "ue" -> "u" and "ü" -> "u", because some bank institutions
-    // remove the dots "ü" -> "u". So we get "u" == "ü" == "ue".
-    return word.toLowerCase().replaceAll("ä", "a").replaceAll("ae", "a")
-        .replaceAll("ö", "o").replaceAll("oe", "o").replaceAll("ü", "u")
-        .replaceAll("ue", "u").replaceAll("ß", "s").replaceAll("ss", "s");
-  }
-
   public Button getStartKontoauszugButton(final Object currentObject,
       final MitgliedskontoControl control)
   {
@@ -1267,22 +886,17 @@ public class MitgliedskontoControl extends DruckMailControl
     return false;
   }
   
-
-  public Object[] getCVSExportGrenzen(Mitglied selectedMitglied)
+  public Object[] getCVSExportGrenzen() throws RemoteException
   {
-    return new Object[] {
+    return new Object[] 
+      {
+        getSuchname().getValue(),
+        getDifferenz().getValue(),
+        getOhneAbbucher().getValue(),
         getDatumvon().getValue(),
         getDatumbis().getValue(),
-        getDifferenz().getValue(), getCVSExportGrenzeOhneAbbucher(),
-        selectedMitglied };
-  }
-
-
-  private Boolean getCVSExportGrenzeOhneAbbucher()
-  {
-    if (null == ohneabbucher)
-      return Boolean.FALSE;
-    return (Boolean) ohneabbucher.getValue();
+        getMailauswahl().getValue()
+      };
   }
 
 }
